refactor: structural debt remediation — spec iteration 2 (II-1 through II-6)
Six independent structural improvements, no behavior changes:

II-1 — Decompose add/mod.rs (1,856 → 350 lines) into real submodules
  - sources.rs: add_from_git/zip/folder/registry + parse/resolve helpers
  - install.rs: install_via_download/local_path, finish_skill_install, copy_dir_recursive
  - skill_def.rs: create_skill_from_path, read_*_from_toml helpers
  - mod.rs: orchestration types + execute_add only
  - Re-exports preserve public API (create_skill_from_path, copy_dir_recursive)

II-2 — Replace wildcard glob re-exports in core/mod.rs
  - All pub use X::* replaced with explicit named lists
  - Public API is now auditable without visiting every submodule

II-3 — Remove stdlib/tokio/serde re-exports from lib.rs
  - Removes: HashMap, Arc, Mutex, RwLock, Path, PathBuf, Serialize, Deserialize,
    async_trait — callers import from canonical sources

II-4 — Populate sources/ shim files and shrink mod.rs to ≤116 lines
  - sources/manager.rs, model.rs, local.rs, marketplace.rs fully populated
  - sources/mod.rs becomes thin declarations + SourcesError + curated re-exports

II-5 — Fold triple severity-count passes in execute_duplicates
  - Single fold into SeverityCounts struct replaces three filter/count passes

II-6 — Extract service initialization helpers from FastSkillService::new
  - build_storage_backend, build_sandbox, build_validators, build_vector_index_service
    extracted as named private fns; new() is now a sequencing call
