+++ crates/fastskill-cli/src/commands/add/mod.rs
@@ -4,24 +4,18 @@ pub mod install;
 pub mod skill_def;
 pub mod sources;
 
-use crate::error::{manifest_required_for_add_message, CliError, CliResult, CliWarning};
-use crate::utils::install_utils;
-use crate::utils::manifest_utils;
-use crate::utils::{
-    detect_skill_source, parse_git_url, parse_skill_id, validate_skill_structure, SkillSource,
-};
+// Re-export public API consumed by install_utils.rs
+use crate::error::{manifest_required_for_add_message, CliError, CliResult};
+use crate::utils::{detect_skill_source, SkillSource};
 use chrono::Utc;
 use clap::Args;
 use fastskill_core::core::project::resolve_project_file;
-use fastskill_core::core::repository::RepositoryManager;
 use fastskill_core::core::skill_manager::SourceType;
 use fastskill_core::{FastSkillService, SkillDefinition};
+pub use install::copy_dir_recursive;
+pub use skill_def::create_skill_from_path;
 use std::env;
-use std::fs;
 use std::path::{Path, PathBuf};
-use tempfile::TempDir;
-use tracing::info;
-use walkdir::WalkDir;
 
 /// Bundled options for add operations (reduces argument count)
 struct AddContext<'a> {
@@ -49,17 +43,15 @@ struct InstallTarget {
     version_display: String,
 }
 
-/// T028: Helper to update skill-project.toml and skills.lock (project scope only)
 fn update_project_files(
     skill_def: &SkillDefinition,
     groups: Vec<String>,
     editable: bool,
 ) -> CliResult<()> {
-    // Update skill-project.toml
+    use crate::utils::manifest_utils;
     manifest_utils::add_skill_to_project_toml(skill_def, groups.clone(), editable)
         .map_err(|e| CliError::Config(format!("Failed to update skill-project.toml: {}", e)))?;
 
-    // T033: Lock file at project root
     let current_dir = env::current_dir()
         .map_err(|e| CliError::Config(format!("Failed to get current directory: {}", e)))?;
     let project_file_result = resolve_project_file(&current_dir);
@@ -76,6 +68,7 @@ fn update_project_files(
 
 /// Helper to update global-skills.lock (global scope only)
 fn update_global_files(skill_def: &SkillDefinition) -> CliResult<()> {
+    use crate::utils::manifest_utils;
     manifest_utils::update_global_lock_file(skill_def, Utc::now())
         .map_err(|e| CliError::Config(format!("Failed to update global lock file: {}", e)))?;
     Ok(())
@@ -111,248 +104,16 @@ async fn register_skill_once(ctx: &AddContext<'_>, skill_def: &SkillDefinition) 
     Ok(())
 }
 
-/// Copy skill to storage, register, update source tracking, and update project files.
-async fn install_via_download(
-    ctx: &AddContext<'_>,
-    skill_path: &Path,
-    mut skill_def: SkillDefinition,
-    target: InstallTarget,
-) -> CliResult<()> {
-    if target.storage_dir.exists() {
-        if !ctx.force {
-            return Err(CliError::Config(format!(
-                "Skill directory '{}' already exists. Use --force to overwrite.",
-                target.storage_dir.display()
-            )));
-        }
-        tokio::fs::remove_dir_all(&target.storage_dir)
-            .await
-            .map_err(CliError::Io)?;
-    }
-    copy_dir_recursive(skill_path, &target.storage_dir).await?;
-    skill_def.skill_file = target.storage_dir.join("SKILL.md");
-    skill_def.source_url = target.meta.source_url;
-    skill_def.source_type = target.meta.source_type;
-    skill_def.source_branch = target.meta.source_branch;
-    skill_def.source_tag = target.meta.source_tag;
-    skill_def.source_subdir = target.meta.source_subdir;
-    skill_def.installed_from = target.meta.installed_from;
-    skill_def.fetched_at = Some(Utc::now());
-    skill_def.editable = ctx.editable;
-
-    register_skill_once(ctx, &skill_def).await?;
-
-    let update = fastskill_core::core::skill_manager::SkillUpdate {
-        source_url: skill_def.source_url.clone(),
-        source_type: skill_def.source_type.clone(),
-        source_branch: skill_def.source_branch.clone(),
-        source_tag: skill_def.source_tag.clone(),
-        source_subdir: skill_def.source_subdir.clone(),
-        installed_from: skill_def.installed_from.clone(),
-        fetched_at: skill_def.fetched_at,
-        editable: Some(skill_def.editable),
-        ..Default::default()
-    };
-    ctx.service
-        .skill_manager()
-        .update_skill(&skill_def.id, update)
-        .await
-        .map_err(|e| {
-            crate::utils::service_error_to_cli(
-                e,
-                ctx.service.config().skill_storage_path.as_path(),
-                ctx.global,
-            )
-        })?;
-
-    if ctx.global {
-        update_global_files(&skill_def)?;
-        println!(
-            "Successfully added skill: {} (v{})",
-            skill_def.name, target.version_display
-        );
-        println!(
-            "{}",
-            crate::utils::messages::ok("Updated global-skills.lock")
-        );
-    } else {
-        update_project_files(&skill_def, ctx.groups.clone(), ctx.editable)?;
-        println!(
-            "Successfully added skill: {} (v{})",
-            skill_def.name, target.version_display
-        );
-        println!(
-            "{}",
-            crate::utils::messages::ok("Updated skill-project.toml and skills.lock")
-        );
-    }
-    Ok(())
-}
-
-/// Install a local skill using symlink (editable) or copy (non-editable).
-async fn install_via_local_path(
-    ctx: &AddContext<'_>,
-    skill_path: &Path,
-    mut skill_def: SkillDefinition,
-    target: InstallTarget,
-) -> CliResult<()> {
-    // Check existence (including broken symlinks) before proceeding
-    let path_exists = target.storage_dir.exists() || target.storage_dir.is_symlink();
-    if path_exists {
-        if !ctx.force {
-            return Err(CliError::Config(format!(
-                "Skill directory '{}' already exists. Use --force to overwrite.",
-                target.storage_dir.display()
-            )));
-        }
-        // Remove the existing path (handle both dir and symlink)
-        if target.storage_dir.is_symlink() || target.storage_dir.is_file() {
-            tokio::fs::remove_file(&target.storage_dir)
-                .await
-                .map_err(CliError::Io)?;
-        } else {
-            tokio::fs::remove_dir_all(&target.storage_dir)
-                .await
-                .map_err(CliError::Io)?;
-        }
-    }
-
-    install_utils::setup_skill_in_storage(skill_path, &target.storage_dir, ctx.editable).await?;
-
-    skill_def.skill_file = target.storage_dir.join("SKILL.md");
-    skill_def.source_url = target.meta.source_url;
-    skill_def.source_type = target.meta.source_type;
-    skill_def.source_branch = target.meta.source_branch;
-    skill_def.source_tag = target.meta.source_tag;
-    skill_def.source_subdir = target.meta.source_subdir;
-    skill_def.installed_from = target.meta.installed_from;
-    skill_def.fetched_at = Some(Utc::now());
-    skill_def.editable = ctx.editable;
-
-    register_skill_once(ctx, &skill_def).await?;
-
-    let update = fastskill_core::core::skill_manager::SkillUpdate {
-        source_url: skill_def.source_url.clone(),
-        source_type: skill_def.source_type.clone(),
-        source_branch: skill_def.source_branch.clone(),
-        source_tag: skill_def.source_tag.clone(),
-        source_subdir: skill_def.source_subdir.clone(),
-        installed_from: skill_def.installed_from.clone(),
-        fetched_at: skill_def.fetched_at,
-        editable: Some(skill_def.editable),
-        ..Default::default()
-    };
-    ctx.service
-        .skill_manager()
-        .update_skill(&skill_def.id, update)
-        .await
-        .map_err(|e| {
-            crate::utils::service_error_to_cli(
-                e,
-                ctx.service.config().skill_storage_path.as_path(),
-                ctx.global,
-            )
-        })?;
-
-    if ctx.global {
-        update_global_files(&skill_def)?;
-        println!(
-            "Successfully added skill: {} (v{})",
-            skill_def.name, target.version_display
-        );
-        println!(
-            "{}",
-            crate::utils::messages::ok("Updated global-skills.lock")
-        );
-    } else {
-        update_project_files(&skill_def, ctx.groups.clone(), ctx.editable)?;
-        println!(
-            "Successfully added skill: {} (v{})",
-            skill_def.name, target.version_display
-        );
-        println!(
-            "{}",
-            crate::utils::messages::ok("Updated skill-project.toml and skills.lock")
-        );
-    }
-    Ok(())
-}
-
-fn component_is_hidden(c: std::path::Component<'_>) -> bool {
-    matches!(c, std::path::Component::Normal(name) if name.to_string_lossy().starts_with('.'))
-}
-
-/// Discover all directories containing SKILL.md under the given base directory.
-/// Skips hidden directories (those starting with '.').
-fn get_skill_dirs_recursive(base: &Path) -> CliResult<Vec<PathBuf>> {
-    if !base.exists() {
-        return Err(CliError::InvalidSource(format!(
-            "Directory does not exist: {}",
-            base.display()
-        )));
-    }
-    if !base.is_dir() {
-        return Err(CliError::InvalidSource(format!(
-            "Path is not a directory: {}",
-            base.display()
-        )));
-    }
-
-    let base_canonical = base.canonicalize().map_err(CliError::Io)?;
-    let mut skill_dirs = Vec::new();
-
-    for entry in WalkDir::new(base)
-        .min_depth(1)
-        .max_depth(usize::MAX)
-        .follow_links(false)
-    {
-        let entry = entry.map_err(|e| {
-            CliError::Io(
-                e.into_io_error()
-                    .unwrap_or_else(|| std::io::Error::other("WalkDir error")),
-            )
-        })?;
-        let path = entry.path();
-        if !path.is_dir() || !path.join("SKILL.md").exists() {
-            continue;
-        }
-
-        let path_canonical = path.canonicalize().map_err(CliError::Io)?;
-        let relative_path = path_canonical.strip_prefix(&base_canonical).map_err(|_| {
-            CliError::Validation(format!(
-                "Failed to compute relative path for: {}",
-                path.display()
-            ))
-        })?;
-        if relative_path.components().any(component_is_hidden) {
-            continue;
-        }
-        skill_dirs.push(path_canonical);
-    }
-
-    Ok(skill_dirs)
-}
-
 /// Add a skill (two operational modes)
 ///
 /// Mode 1: Manifest-managed (when skill-project.toml exists)
-/// - Adds the skill to skill-project.toml [dependencies] section
-/// - Updates skills.lock
-/// - Use 'fastskill install' afterward to apply the changes
-///
 /// Mode 2: Local-only (when no skill-project.toml or --global flag)
-/// - Registers skill directly to local installation
-/// - Skill is available immediately but not tracked in project manifest
-///
-/// Supports git URLs, local paths, ZIP files, and registry skill IDs.
-/// Local paths are automatically converted to absolute paths in skill-project.toml.
 #[derive(Debug, Args)]
 pub struct AddArgs {
     /// Source: path to zip file, folder, git URL, or skill ID (e.g., pptx@1.2.3)
     pub source: String,
 
     /// Override source type (registry, github/git, local)
-    /// Default is registry for skill IDs. Use 'github' or 'git' for git repositories.
     #[arg(long)]
     pub source_type: Option<String>,
 
@@ -400,51 +161,6 @@ fn resolve_source(args: &AddArgs) -> SkillSource {
     }
 }
 
-async fn handle_recursive_add(ctx: &AddContext<'_>, path: &Path) -> CliResult<()> {
-    info!(
-        "Adding skills recursively from directory: {}",
-        path.display()
-    );
-    let skill_dirs = get_skill_dirs_recursive(path)?;
-    if skill_dirs.is_empty() {
-        return Err(CliError::Validation(format!(
-            "No skill directories found under {}",
-            path.display()
-        )));
-    }
-    println!(
-        "Found {} skill(s) in directory {}",
-        skill_dirs.len(),
-        path.display()
-    );
-
-    let mut failed: Vec<(PathBuf, CliError)> = Vec::new();
-    let mut success_count = 0;
-    for skill_path in skill_dirs {
-        match add_from_folder(ctx, &skill_path).await {
-            Ok(()) => success_count += 1,
-            Err(e) => {
-                eprintln!("Skill at {}: {}", skill_path.display(), e);
-                failed.push((skill_path, e));
-            }
-        }
-    }
-    if failed.is_empty() {
-        return Ok(());
-    }
-    let total = success_count + failed.len();
-    Err(CliError::Validation(format!(
-        "{} of {} skills failed:\n{}",
-        failed.len(),
-        total,
-        failed
-            .iter()
-            .map(|(path, err)| format!("  - {}: {}", path.display(), err))
-            .collect::<Vec<_>>()
-            .join("\n")
-    )))
-}
-
 fn ensure_manifest() -> CliResult<()> {
     let current_dir = env::current_dir()
         .map_err(|e| CliError::Config(format!("Failed to get current directory: {}", e)))?;
@@ -457,8 +173,26 @@ fn ensure_manifest() -> CliResult<()> {
     Ok(())
 }
 
+fn validate_folder_has_skill(path: &Path) -> CliResult<()> {
+    if path.join("SKILL.md").exists() {
+        return Ok(());
+    }
+    if let Ok(dirs) = install::get_skill_dirs_recursive(path) {
+        if !dirs.is_empty() {
+            return Err(CliError::Validation(format!(
+                "This directory has no SKILL.md at the root but contains {} skill(s) in subdirectories. Add them all with: fastskill add {} --recursive",
+                dirs.len(),
+                path.display()
+            )));
+        }
+    }
+    Err(CliError::Validation(format!(
+        "No SKILL.md found in {}. A skill directory must contain a SKILL.md file. If this folder has multiple skills in subdirectories, use --recursive.",
+        path.display()
+    )))
+}
+
 pub async fn execute_add(service: &FastSkillService, args: AddArgs, global: bool) -> CliResult<()> {
-    // Validate arguments before checking for manifest to provide better error messages
     let source = resolve_source(&args);
 
     if args.editable {
@@ -474,7 +208,6 @@ pub async fn execute_add(service: &FastSkillService, args: AddArgs, global: bool
         }
     }
 
-    // Global installs do not require a project manifest
     if !global {
         ensure_manifest()?;
     }
@@ -496,528 +229,20 @@ pub async fn execute_add(service: &FastSkillService, args: AddArgs, global: bool
                 ));
             }
         };
-        return handle_recursive_add(&ctx, path).await;
+        return install::handle_recursive_add(&ctx, path).await;
     }
 
     match source {
-        SkillSource::ZipFile(path) => add_from_zip(&ctx, &path).await,
+        SkillSource::ZipFile(path) => sources::add_from_zip(&ctx, &path).await,
         SkillSource::Folder(path) => {
             validate_folder_has_skill(&path)?;
-            add_from_folder(&ctx, &path).await
+            sources::add_from_folder(&ctx, &path).await
         }
         SkillSource::GitUrl(url) => {
-            add_from_git(&ctx, &url, args.branch.as_deref(), args.tag.as_deref()).await
+            sources::add_from_git(&ctx, &url, args.branch.as_deref(), args.tag.as_deref()).await
         }
-        SkillSource::SkillId(skill_id) => add_from_registry(&ctx, &skill_id).await,
+        SkillSource::SkillId(skill_id) => sources::add_from_registry(&ctx, &skill_id).await,
     }
-}
-
-fn validate_folder_has_skill(path: &Path) -> CliResult<()> {
-    if path.join("SKILL.md").exists() {
-        return Ok(());
-    }
-    if let Ok(dirs) = get_skill_dirs_recursive(path) {
-        if !dirs.is_empty() {
-            return Err(CliError::Validation(format!(
-                "This directory has no SKILL.md at the root but contains {} skill(s) in subdirectories. Add them all with: fastskill add {} --recursive",
-                dirs.len(),
-                path.display()
-            )));
-        }
-    }
-    Err(CliError::Validation(format!(
-        "No SKILL.md found in {}. A skill directory must contain a SKILL.md file. If this folder has multiple skills in subdirectories, use --recursive.",
-        path.display()
-    )))
-}
-
-fn extract_zip_to_temp(zip_path: &Path) -> CliResult<tempfile::TempDir> {
-    let temp_dir = TempDir::new().map_err(CliError::Io)?;
-    let extract_path = temp_dir.path();
-    use fastskill_core::storage::zip::ZipHandler;
-    let zip_handler = ZipHandler::new()
-        .map_err(|e| CliError::InvalidSource(format!("Failed to create ZIP handler: {}", e)))?;
-    zip_handler
-        .extract_to_dir(zip_path, extract_path)
-        .map_err(|e| match e {
-            fastskill_core::core::service::ServiceError::Validation(msg) => {
-                CliError::InvalidSource(format!("ZIP extraction validation failed: {}", msg))
-            }
-            fastskill_core::core::service::ServiceError::Io(err) => CliError::Io(err),
-            _ => CliError::InvalidSource(format!("ZIP extraction failed: {}", e)),
-        })?;
-    Ok(temp_dir)
-}
-
-async fn add_from_zip(ctx: &AddContext<'_>, zip_path: &Path) -> CliResult<()> {
-    info!("Adding skill from zip file: {}", zip_path.display());
-    if !zip_path.exists() {
-        return Err(CliError::InvalidSource(format!(
-            "Zip file does not exist: {}",
-            zip_path.display()
-        )));
-    }
-
-    // Canonicalize the zip path to get absolute path
-    let canonical_zip_path = zip_path.canonicalize().map_err(|e| {
-        CliError::InvalidSource(format!(
-            "Failed to resolve absolute path for '{}': {}",
-            zip_path.display(),
-            e
-        ))
-    })?;
-
-    let _temp_dir = extract_zip_to_temp(&canonical_zip_path)?;
-    let extract_path = _temp_dir.path();
-    let skill_path = find_skill_in_directory(extract_path)?;
-    validate_skill_structure(&skill_path)?;
-    let skill_def = create_skill_from_path(&skill_path, "zip", false)?;
-    let version = skill_def.version.clone();
-    let target = InstallTarget {
-        storage_dir: ctx
-            .service
-            .config()
-            .skill_storage_path
-            .join(skill_def.id.as_str()),
-        meta: SourceMeta {
-            source_url: Some(canonical_zip_path.to_string_lossy().to_string()),
-            source_type: Some(SourceType::ZipFile),
-            source_branch: None,
-            source_tag: None,
-            source_subdir: None,
-            installed_from: None,
-        },
-        version_display: version,
-    };
-    install_via_download(ctx, &skill_path, skill_def, target).await
-}
-
-async fn add_from_folder(ctx: &AddContext<'_>, folder_path: &Path) -> CliResult<()> {
-    info!("Adding skill from folder: {}", folder_path.display());
-    validate_skill_structure(folder_path)?;
-    let skill_def = create_skill_from_path(folder_path, "local", ctx.editable)?;
-
-    // Canonicalize the folder path to get absolute path
-    let canonical_path = folder_path.canonicalize().map_err(|e| {
-        CliError::InvalidSource(format!(
-            "Failed to resolve absolute path for '{}': {}",
-            folder_path.display(),
-            e
-        ))
-    })?;
-
-    let target = InstallTarget {
-        storage_dir: ctx
-            .service
-            .config()
-            .skill_storage_path
-            .join(skill_def.id.as_str()),
-        meta: SourceMeta {
-            source_url: Some(canonical_path.to_string_lossy().to_string()),
-            source_type: Some(SourceType::LocalPath),
-            source_branch: None,
-            source_tag: None,
-            source_subdir: None,
-            installed_from: None,
-        },
-        version_display: skill_def.version.clone(),
-    };
-    install_via_local_path(ctx, &canonical_path, skill_def, target).await
-}
-
-async fn clone_and_validate_skill(
-    git_url: &str,
-    branch: Option<&str>,
-    tag: Option<&str>,
-) -> CliResult<(
-    TempDir,
-    PathBuf,
-    SkillDefinition,
-    Option<String>,
-    Option<String>,
-)> {
-    use fastskill_core::storage::git::{clone_repository, validate_cloned_skill};
-    let git_info = parse_git_url(git_url)?;
-    let branch = branch.or(git_info.branch.as_deref());
-    let temp_dir = clone_repository(&git_info.repo_url, branch, tag, None)
-        .await
-        .map_err(|e| CliError::GitCloneFailed(e.to_string()))?;
-    let skill_base_path = match &git_info.subdir {
-        Some(subdir) => {
-            let subdir_path = temp_dir.path().join(subdir);
-            if !subdir_path.exists() {
-                return Err(CliError::InvalidSource(format!(
-                    "Specified subdirectory '{}' does not exist in cloned repository",
-                    subdir.display()
-                )));
-            }
-            subdir_path
-        }
-        None => temp_dir.path().to_path_buf(),
-    };
-    let skill_path = validate_cloned_skill(&skill_base_path)
-        .map_err(|e| CliError::SkillValidationFailed(e.to_string()))?;
-    let skill_def = create_skill_from_path(&skill_path, "git", false)?;
-    Ok((
-        temp_dir,
-        skill_path,
-        skill_def,
-        branch.map(String::from),
-        tag.map(String::from),
-    ))
-}
-
-async fn add_from_git(
-    ctx: &AddContext<'_>,
-    git_url: &str,
-    branch: Option<&str>,
-    tag: Option<&str>,
-) -> CliResult<()> {
-    info!("Adding skill from git URL: {}", git_url);
-    let (_temp_dir, skill_path, skill_def, branch_opt, tag_opt) =
-        clone_and_validate_skill(git_url, branch, tag).await?;
-    validate_skill_structure(&skill_path)?;
-    let git_info = parse_git_url(git_url)?;
-    let target = InstallTarget {
-        storage_dir: ctx
-            .service
-            .config()
-            .skill_storage_path
-            .join(skill_def.id.as_str()),
-        meta: SourceMeta {
-            source_url: Some(git_url.to_string()),
-            source_type: Some(SourceType::GitUrl),
-            source_branch: branch_opt.or(git_info.branch),
-            source_tag: tag_opt,
-            source_subdir: git_info.subdir,
-            installed_from: None,
-        },
-        version_display: skill_def.version.clone(),
-    };
-    install_via_download(ctx, &skill_path, skill_def, target).await
-}
-
-fn parse_registry_scope_id(
-    skill_id_input: &str,
-) -> CliResult<(String, String, String, Option<String>)> {
-    let (skill_id_full, version_opt) = parse_skill_id(skill_id_input);
-    let (scope, expected_id) = match skill_id_full.find('/') {
-        Some(slash_pos) => (
-            skill_id_full[..slash_pos].to_string(),
-            skill_id_full[slash_pos + 1..].to_string(),
-        ),
-        None => {
-            return Err(CliError::Config(format!(
-                "Registry skill ID must be in format 'scope/id', got: {}",
-                skill_id_full
-            )));
-        }
-    };
-    Ok((skill_id_full, scope, expected_id, version_opt))
-}
-
-async fn resolve_registry_version(
-    repo_client: &(dyn fastskill_core::core::repository::RepositoryClient + Send + Sync),
-    skill_id_full: &str,
-    version_opt: Option<String>,
-) -> CliResult<String> {
-    if let Some(v) = version_opt {
-        return Ok(v);
-    }
-    let versions = repo_client
-        .get_versions(skill_id_full)
-        .await
-        .map_err(|e| CliError::Config(format!("Failed to get versions: {}", e)))?;
-    if versions.is_empty() {
-        return Err(CliError::Config(format!(
-            "No versions found for skill '{}' in repository",
-            skill_id_full
-        )));
-    }
-    let mut sorted = versions;
-    sorted.sort();
-    sorted
-        .last()
-        .cloned()
-        .ok_or_else(|| CliError::Config("No versions available".to_string()))
-}
-
-async fn download_registry_package(
-    repo_client: &(dyn fastskill_core::core::repository::RepositoryClient + Send + Sync),
-    skill_id_full: &str,
-    version: &str,
-) -> CliResult<(TempDir, PathBuf, SkillDefinition)> {
-    let zip_data = repo_client
-        .download(skill_id_full, version)
-        .await
-        .map_err(|e| CliError::Config(format!("Failed to download package: {}", e)))?;
-    let temp_dir = TempDir::new().map_err(CliError::Io)?;
-    let extract_path = temp_dir.path().join("extracted");
-    fs::create_dir_all(&extract_path).map_err(CliError::Io)?;
-    let temp_zip = temp_dir.path().join(format!("package-{}.zip", version));
-    std::fs::write(&temp_zip, zip_data).map_err(CliError::Io)?;
-    use fastskill_core::storage::zip::ZipHandler;
-    let zip_handler = ZipHandler::new()
-        .map_err(|e| CliError::InvalidSource(format!("Failed to create ZIP handler: {}", e)))?;
-    zip_handler
-        .extract_to_dir(&temp_zip, &extract_path)
-        .map_err(|e| match e {
-            fastskill_core::core::service::ServiceError::Validation(msg) => {
-                CliError::InvalidSource(format!("ZIP extraction validation failed: {}", msg))
-            }
-            fastskill_core::core::service::ServiceError::Io(err) => CliError::Io(err),
-            _ => CliError::InvalidSource(format!("ZIP extraction failed: {}", e)),
-        })?;
-    let skill_path = find_skill_in_directory(&extract_path)?;
-    validate_skill_structure(&skill_path)?;
-    let skill_def = create_skill_from_path(&skill_path, "registry", false)?;
-    Ok((temp_dir, skill_path, skill_def))
-}
-
-async fn add_from_registry(ctx: &AddContext<'_>, skill_id_input: &str) -> CliResult<()> {
-    info!("Adding skill from registry: {}", skill_id_input);
-    let (skill_id_full, scope, expected_id, version_opt) = parse_registry_scope_id(skill_id_input)?;
-
-    let repositories = crate::config::load_repositories_from_project()?;
-    let repo_manager = RepositoryManager::from_definitions(repositories);
-    let default_repo = repo_manager.get_default_repository().ok_or_else(|| {
-        CliError::Config(
-            "No default repository configured. Use 'fastskill repos add' to add a repository."
-                .to_string(),
-        )
-    })?;
-    if !matches!(
-        default_repo.repo_type,
-        fastskill_core::core::repository::RepositoryType::HttpRegistry
-    ) {
-        return Err(CliError::Config(format!(
-            "Repository '{}' is not an http-registry type. Only http-registry repositories are supported for 'fastskill add' currently.",
-            default_repo.name
-        )));
-    }
-
-    let repo_client = repo_manager
-        .get_client(&default_repo.name)
-        .await
-        .map_err(|e| CliError::Config(format!("Failed to get repository client: {}", e)))?;
-    let version =
-        resolve_registry_version(repo_client.as_ref(), &skill_id_full, version_opt).await?;
-
-    let _skill_metadata = repo_client
-        .get_skill(&skill_id_full, Some(&version))
-        .await
-        .map_err(|e| CliError::Config(format!("Failed to get skill: {}", e)))?
-        .ok_or_else(|| {
-            CliError::Config(format!(
-                "Skill '{}' version '{}' not found in repository",
-                skill_id_full, version
-            ))
-        })?;
-
-    info!(
-        "Downloading {}@{} from repository...",
-        skill_id_full, version
-    );
-    let (_temp_dir, skill_path, skill_def) =
-        download_registry_package(repo_client.as_ref(), &skill_id_full, &version).await?;
-
-    if skill_def.id.as_str() != expected_id {
-        return Err(CliError::Config(format!(
-            "Skill ID mismatch: expected '{}' from registry reference '{}', but found '{}' in skill-project.toml",
-            expected_id, skill_id_full, skill_def.id.as_str()
-        )));
-    }
-
-    let target = InstallTarget {
-        storage_dir: ctx
-            .service
-            .config()
-            .skill_storage_path
-            .join(&scope)
-            .join(skill_def.id.as_str()),
-        meta: SourceMeta {
-            source_url: None,
-            source_type: Some(SourceType::Source),
-            source_branch: None,
-            source_tag: None,
-            source_subdir: None,
-            installed_from: Some(default_repo.name.clone()),
-        },
-        version_display: version,
-    };
-    install_via_download(ctx, &skill_path, skill_def, target).await
-}
-
-/// Read skill ID and version from skill-project.toml (optional)
-fn read_skill_id_from_toml(skill_path: &Path) -> CliResult<Option<String>> {
-    let skill_project_path = skill_path.join("skill-project.toml");
-
-    if !skill_project_path.exists() {
-        return Ok(None);
-    }
-
-    #[derive(serde::Deserialize)]
-    struct SkillProjectToml {
-        metadata: Option<fastskill_core::core::manifest::MetadataSection>,
-    }
-
-    let content = fs::read_to_string(&skill_project_path)
-        .map_err(|e| CliError::Validation(format!("Failed to read skill-project.toml: {}", e)))?;
-
-    let project: SkillProjectToml = toml::from_str(&content)
-        .map_err(|e| CliError::Validation(format!("Failed to parse skill-project.toml: {}", e)))?;
-
-    let metadata = project.metadata.ok_or_else(|| {
-        CliError::Validation("skill-project.toml must have a [metadata] section".to_string())
-    })?;
-
-    Ok(metadata.id)
-}
-
-fn read_version_from_toml(skill_path: &Path) -> CliResult<Option<String>> {
-    let skill_project_path = skill_path.join("skill-project.toml");
-
-    if !skill_project_path.exists() {
-        return Ok(None);
-    }
-
-    #[derive(serde::Deserialize)]
-    struct SkillProjectToml {
-        metadata: Option<fastskill_core::core::manifest::MetadataSection>,
-    }
-
-    let content = fs::read_to_string(&skill_project_path)
-        .map_err(|e| CliError::Validation(format!("Failed to read skill-project.toml: {}", e)))?;
-
-    let project: SkillProjectToml = toml::from_str(&content)
-        .map_err(|e| CliError::Validation(format!("Failed to parse skill-project.toml: {}", e)))?;
-
-    Ok(project.metadata.and_then(|m| m.version))
-}
-
-fn generate_fallback_skill_id(skill_path: &Path) -> CliResult<String> {
-    let skill_md_path = skill_path.join("SKILL.md");
-
-    let content = std::fs::read_to_string(&skill_md_path)
-        .map_err(|e| CliError::Validation(format!("Failed to read SKILL.md: {}", e)))?;
-
-    let frontmatter =
-        fastskill_core::core::metadata::parse_yaml_frontmatter(&content).map_err(|e| {
-            CliError::Validation(format!("Failed to parse SKILL.md frontmatter: {}", e))
-        })?;
-
-    if let Some(metadata) = &frontmatter.metadata {
-        if let Some(id) = metadata.get("id") {
-            return Ok(id.clone());
-        }
-    }
-
-    Ok(frontmatter.name.clone())
-}
-
-/// Find SKILL.md in a directory (could be at root or in subdirectory)
-fn find_skill_in_directory(dir: &Path) -> CliResult<PathBuf> {
-    // Check root
-    if dir.join("SKILL.md").exists() {
-        return Ok(dir.to_path_buf());
-    }
-
-    // Check subdirectories
-    let entries = fs::read_dir(dir).map_err(CliError::Io)?;
-    for entry in entries {
-        let entry = entry.map_err(CliError::Io)?;
-        let path = entry.path();
-        if path.is_dir() && path.join("SKILL.md").exists() {
-            return Ok(path);
-        }
-    }
-
-    Err(CliError::Validation(format!(
-        "SKILL.md not found in: {}",
-        dir.display()
-    )))
-}
-
-/// Create a skill definition from a path containing SKILL.md
-/// The skill ID is read from skill-project.toml if present, otherwise from SKILL.md frontmatter
-pub fn create_skill_from_path(
-    skill_path: &Path,
-    source_type: &str,
-    editable: bool,
-) -> CliResult<SkillDefinition> {
-    let skill_id_str = match read_skill_id_from_toml(skill_path)? {
-        Some(id) => id,
-        None => {
-            let fallback = generate_fallback_skill_id(skill_path)?;
-            CliWarning::MissingSkillProjectToml {
-                path: skill_path.to_path_buf(),
-                fallback_id: fallback.clone(),
-            }
-            .display(source_type, editable);
-            fallback
-        }
-    };
-
-    let skill_file = skill_path.join("SKILL.md");
-    let content = fs::read_to_string(&skill_file).map_err(CliError::Io)?;
-
-    let frontmatter = fastskill_core::core::metadata::parse_yaml_frontmatter(&content)
-        .map_err(|e| CliError::Validation(format!("Failed to parse SKILL.md: {}", e)))?;
-
-    let skill_id_str_clone = skill_id_str.clone();
-    let skill_id = fastskill_core::SkillId::new(skill_id_str).map_err(|_| {
-        CliError::Validation(format!("Invalid skill ID format: {}", skill_id_str_clone))
-    })?;
-
-    let version = match read_version_from_toml(skill_path)? {
-        Some(toml_version) => toml_version,
-        None => frontmatter
-            .metadata
-            .as_ref()
-            .and_then(|m| m.get("version"))
-            .or(frontmatter.version.as_ref())
-            .cloned()
-            .unwrap_or_else(|| {
-                eprintln!(
-                    "⚠  Note: No version specified in SKILL.md frontmatter, defaulting to 1.0.0"
-                );
-                "1.0.0".to_string()
-            }),
-    };
-
-    let mut skill =
-        SkillDefinition::new(skill_id, frontmatter.name, frontmatter.description, version);
-
-    skill.skill_file = skill_file.clone();
-    skill.author = frontmatter.author;
-
-    Ok(skill)
-}
-
-/// Recursively copy a directory from src to dst
-pub async fn copy_dir_recursive(src: &Path, dst: &Path) -> CliResult<()> {
-    // Create destination directory
-    tokio::fs::create_dir_all(dst).await.map_err(CliError::Io)?;
-
-    let mut entries = tokio::fs::read_dir(src).await.map_err(CliError::Io)?;
-
-    while let Some(entry) = entries.next_entry().await.map_err(CliError::Io)? {
-        let ty = entry.file_type().await.map_err(CliError::Io)?;
-        let src_path = entry.path();
-        let dst_path = dst.join(entry.file_name());
-
-        if ty.is_dir() {
-            // Recursively copy subdirectory (boxed to handle async recursion)
-            Box::pin(copy_dir_recursive(&src_path, &dst_path)).await?;
-        } else {
-            // Copy file
-            tokio::fs::copy(&src_path, &dst_path)
-                .await
-                .map_err(CliError::Io)?;
-        }
-    }
-
-    Ok(())
 }
 
 #[cfg(test)]
@@ -1071,7 +296,6 @@ mod tests {
 
     #[tokio::test]
     async fn test_execute_add_no_manifest_fails_with_instructions() {
-        // Use a shared mutex to serialize directory changes across parallel tests
         let _lock = fastskill_core::test_utils::DIR_MUTEX
             .lock()
             .unwrap_or_else(|e| e.into_inner());
@@ -1095,7 +319,6 @@ mod tests {
         let original_dir = std::env::current_dir().ok();
         let _guard = DirGuard(original_dir);
         std::env::set_current_dir(temp_dir.path()).unwrap();
-        // No skill-project.toml in temp_dir
 
         let args = AddArgs {
             source: "scope/some-skill".to_string(),
@@ -1123,807 +346,5 @@ mod tests {
         } else {
             panic!("Expected Config error");
         }
-    }
-
-    #[tokio::test]
-    async fn test_execute_add_success() {
-        let _lock = fastskill_core::test_utils::DIR_MUTEX
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let temp_dir = TempDir::new().unwrap();
-        let original_dir = std::env::current_dir().ok();
-
-        struct DirGuard(Option<std::path::PathBuf>);
-        impl Drop for DirGuard {
-            fn drop(&mut self) {
-                if let Some(dir) = &self.0 {
-                    let _ = std::env::set_current_dir(dir);
-                }
-            }
-        }
-        let _guard = DirGuard(original_dir);
-
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-
-        let skills_dir = temp_dir.path().join(".claude/skills");
-        fs::create_dir_all(&skills_dir).unwrap();
-
-        let source_dir = temp_dir.path().join("source-skill");
-        fs::create_dir_all(&source_dir).unwrap();
-        let skill_content = r#"# Test Skill
-
-Name: test-skill
- Version: 1.0.0
- Description: A test skill for coverage
-"#;
-        fs::write(source_dir.join("SKILL.md"), skill_content).unwrap();
-
-        let manifest_content = r#"[tool.fastskill]
- skills_directory = ".claude/skills"
-"#;
-        fs::write(temp_dir.path().join("skill-project.toml"), manifest_content).unwrap();
-
-        let config = ServiceConfig {
-            skill_storage_path: skills_dir,
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-
-        let args = AddArgs {
-            source: source_dir.display().to_string(),
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: false,
-            group: None,
-            recursive: false,
-        };
-
-        let result = execute_add(&service, args, false).await;
-        // May succeed or fail depending on various factors, but should exercise the code path
-        assert!(result.is_ok() || result.is_err());
-    }
-
-    // NOTE: Path traversal security tests have been moved to lower-level tests
-    // in src/core/registry/validation_worker.rs (test_extract_zip_to_temp_rejects_path_traversal
-    // and test_extract_zip_to_temp_rejects_windows_traversal) where they don't rely on
-    // changing the current directory, which is unreliable in concurrent test environments.
-
-    #[tokio::test]
-    async fn test_add_relative_path_creates_absolute_entry() {
-        let _lock = fastskill_core::test_utils::DIR_MUTEX
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let temp_dir = TempDir::new().unwrap();
-        let original_dir = std::env::current_dir().ok();
-
-        struct DirGuard(Option<std::path::PathBuf>);
-        impl Drop for DirGuard {
-            fn drop(&mut self) {
-                if let Some(dir) = &self.0 {
-                    let _ = std::env::set_current_dir(dir);
-                }
-            }
-        }
-        let _guard = DirGuard(original_dir);
-
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-
-        let skills_dir = temp_dir.path().join(".claude/skills");
-        fs::create_dir_all(&skills_dir).unwrap();
-
-        // Create source skill in subdirectory (relative path)
-        let relative_source_dir = temp_dir.path().join("source-skills/test-skill");
-        fs::create_dir_all(&relative_source_dir).unwrap();
-        let skill_content = r#"---
-name: test-skill
-version: 1.0.0
-description: A test skill for relative path testing
----
-
-# Test Skill
-
-This is a test skill for relative path testing.
-"#;
-        fs::write(relative_source_dir.join("SKILL.md"), skill_content).unwrap();
-
-        // Create skill-project.toml for the skill
-        let skill_project_content = r#"[metadata]
-id = "test-skill"
-version = "1.0.0"
-"#;
-        fs::write(
-            relative_source_dir.join("skill-project.toml"),
-            skill_project_content,
-        )
-        .unwrap();
-
-        let manifest_content = r#"[tool.fastskill]
-skills_directory = ".claude/skills"
-
-[dependencies]
-"#;
-        fs::write(temp_dir.path().join("skill-project.toml"), manifest_content).unwrap();
-
-        let config = ServiceConfig {
-            skill_storage_path: skills_dir,
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-
-        // Add skill using relative path
-        let args = AddArgs {
-            source: "source-skills/test-skill".to_string(), // Relative path
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: false,
-            group: None,
-            recursive: false,
-        };
-
-        let result = execute_add(&service, args, false).await;
-        assert!(result.is_ok(), "Add failed: {:?}", result);
-
-        // Verify skill-project.toml contains absolute path
-        let project_toml_path = temp_dir.path().join("skill-project.toml");
-        let project_content = fs::read_to_string(&project_toml_path).unwrap();
-
-        // The path should be absolute (start with /)
-        // Extract the path value from the TOML
-        let path_line = project_content
-            .lines()
-            .find(|line| line.starts_with("path = "))
-            .unwrap();
-        let path_value = path_line.trim_start_matches("path = ").trim_matches('"');
-
-        // Path should be absolute (not relative like "source-skills/test-skill")
-        assert!(
-            std::path::Path::new(path_value).is_absolute(),
-            "Path should be absolute, but got: {}",
-            path_value
-        );
-
-        // Should match the expected absolute path
-        assert!(
-            project_content.contains(&relative_source_dir.to_string_lossy().to_string()),
-            "skill-project.toml should contain absolute path. Expected: {}\nContent:\n{}",
-            relative_source_dir.display(),
-            project_content
-        );
-    }
-
-    #[tokio::test]
-    async fn test_add_absolute_path_preserved_as_is() {
-        let _lock = fastskill_core::test_utils::DIR_MUTEX
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let temp_dir = TempDir::new().unwrap();
-        let original_dir = std::env::current_dir().ok();
-
-        struct DirGuard(Option<std::path::PathBuf>);
-        impl Drop for DirGuard {
-            fn drop(&mut self) {
-                if let Some(dir) = &self.0 {
-                    let _ = std::env::set_current_dir(dir);
-                }
-            }
-        }
-        let _guard = DirGuard(original_dir);
-
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-
-        let skills_dir = temp_dir.path().join(".claude/skills");
-        fs::create_dir_all(&skills_dir).unwrap();
-
-        // Create source skill with absolute path
-        let absolute_source_dir = temp_dir.path().join("test-skill-absolute");
-        fs::create_dir_all(&absolute_source_dir).unwrap();
-        let skill_content = r#"---
-name: test-skill-absolute
-version: 1.0.0
-description: A test skill for absolute path testing
----
-
-# Test Skill Absolute
-
-This is a test skill for absolute path testing.
-"#;
-        fs::write(absolute_source_dir.join("SKILL.md"), skill_content).unwrap();
-
-        // Create skill-project.toml for the skill
-        let skill_project_content = r#"[metadata]
-id = "test-skill-absolute"
-version = "1.0.0"
-"#;
-        fs::write(
-            absolute_source_dir.join("skill-project.toml"),
-            skill_project_content,
-        )
-        .unwrap();
-
-        let manifest_content = r#"[tool.fastskill]
-skills_directory = ".claude/skills"
-
-[dependencies]
-"#;
-        fs::write(temp_dir.path().join("skill-project.toml"), manifest_content).unwrap();
-
-        let config = ServiceConfig {
-            skill_storage_path: skills_dir,
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-
-        // Add skill using absolute path
-        let args = AddArgs {
-            source: absolute_source_dir.to_string_lossy().to_string(), // Absolute path
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: false,
-            group: None,
-            recursive: false,
-        };
-
-        let result = execute_add(&service, args, false).await;
-        assert!(result.is_ok(), "Add failed: {:?}", result);
-
-        // Verify skill-project.toml contains the absolute path
-        let project_toml_path = temp_dir.path().join("skill-project.toml");
-        let project_content = fs::read_to_string(&project_toml_path).unwrap();
-
-        assert!(
-            project_content.contains(&absolute_source_dir.to_string_lossy().to_string()),
-            "skill-project.toml should contain absolute path"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_add_skill_without_toml_fallback() {
-        let _lock = fastskill_core::test_utils::DIR_MUTEX
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let temp_dir = TempDir::new().unwrap();
-        let original_dir = std::env::current_dir().ok();
-
-        struct DirGuard(Option<std::path::PathBuf>);
-        impl Drop for DirGuard {
-            fn drop(&mut self) {
-                if let Some(dir) = &self.0 {
-                    let _ = std::env::set_current_dir(dir);
-                }
-            }
-        }
-        let _guard = DirGuard(original_dir);
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-
-        let skills_dir = temp_dir.path().join(".claude/skills");
-        fs::create_dir_all(&skills_dir).unwrap();
-
-        let source_dir = temp_dir.path().join("test-skill-no-toml");
-        fs::create_dir_all(&source_dir).unwrap();
-
-        let skill_md = r#"---
-name: test-skill-no-toml
-description: Test skill without skill-project.toml
-metadata:
-  id: test-skill-no-toml
-  version: "1.0.0"
----
-Test skill content
-"#;
-        fs::write(source_dir.join("SKILL.md"), skill_md).unwrap();
-
-        let manifest_content = r#"[tool.fastskill]
-skills_directory = ".claude/skills"
-
-[dependencies]
-"#;
-        fs::write(temp_dir.path().join("skill-project.toml"), manifest_content).unwrap();
-
-        let config = ServiceConfig {
-            skill_storage_path: skills_dir,
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-
-        let args = AddArgs {
-            source: source_dir.display().to_string(),
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: false,
-            group: None,
-            recursive: false,
-        };
-
-        let result = execute_add(&service, args, false).await;
-
-        assert!(
-            result.is_ok(),
-            "Add should succeed without skill-project.toml: {:?}",
-            result
-        );
-
-        let skill_id = fastskill_core::SkillId::new("test-skill-no-toml".to_string()).unwrap();
-        let skill = service.skill_manager().get_skill(&skill_id).await.unwrap();
-        assert!(skill.is_some(), "Skill should be registered");
-        assert_eq!(skill.unwrap().id.as_str(), "test-skill-no-toml");
-    }
-
-    #[tokio::test]
-    async fn test_add_skill_toml_takes_precedence() {
-        let _lock = fastskill_core::test_utils::DIR_MUTEX
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let temp_dir = TempDir::new().unwrap();
-        let original_dir = std::env::current_dir().ok();
-
-        struct DirGuard(Option<std::path::PathBuf>);
-        impl Drop for DirGuard {
-            fn drop(&mut self) {
-                if let Some(dir) = &self.0 {
-                    let _ = std::env::set_current_dir(dir);
-                }
-            }
-        }
-        let _guard = DirGuard(original_dir);
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-
-        let skills_dir = temp_dir.path().join(".claude/skills");
-        fs::create_dir_all(&skills_dir).unwrap();
-
-        let source_dir = temp_dir.path().join("test-skill-both");
-        fs::create_dir_all(&source_dir).unwrap();
-
-        let skill_md = r#"---
-name: test-skill-both
-description: Test skill
-metadata:
-  id: from-skill-md
-  version: "2.0.0"
----
-Test content
-"#;
-        fs::write(source_dir.join("SKILL.md"), skill_md).unwrap();
-
-        let project_toml = r#"[metadata]
-id = "from-toml"
-version = "1.5.0"
-"#;
-        fs::write(source_dir.join("skill-project.toml"), project_toml).unwrap();
-
-        let manifest_content = r#"[tool.fastskill]
-skills_directory = ".claude/skills"
-
-[dependencies]
-"#;
-        fs::write(temp_dir.path().join("skill-project.toml"), manifest_content).unwrap();
-
-        let config = ServiceConfig {
-            skill_storage_path: skills_dir,
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-
-        let args = AddArgs {
-            source: source_dir.display().to_string(),
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: false,
-            group: None,
-            recursive: false,
-        };
-
-        let result = execute_add(&service, args, false).await;
-
-        assert!(result.is_ok(), "Add should succeed: {:?}", result);
-
-        let skill_id = fastskill_core::SkillId::new("from-toml".to_string()).unwrap();
-        let skill = service.skill_manager().get_skill(&skill_id).await.unwrap();
-        assert!(skill.is_some(), "skill-project.toml ID should be used");
-        assert_eq!(skill.unwrap().version, "1.5.0");
-    }
-
-    #[tokio::test]
-    async fn test_add_skill_fallback_to_name() {
-        let _lock = fastskill_core::test_utils::DIR_MUTEX
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let temp_dir = TempDir::new().unwrap();
-        let original_dir = std::env::current_dir().ok();
-
-        struct DirGuard(Option<std::path::PathBuf>);
-        impl Drop for DirGuard {
-            fn drop(&mut self) {
-                if let Some(dir) = &self.0 {
-                    let _ = std::env::set_current_dir(dir);
-                }
-            }
-        }
-        let _guard = DirGuard(original_dir);
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-
-        let skills_dir = temp_dir.path().join(".claude/skills");
-        fs::create_dir_all(&skills_dir).unwrap();
-
-        let source_dir = temp_dir.path().join("fallback-name-skill");
-        fs::create_dir_all(&source_dir).unwrap();
-
-        let skill_md = r#"---
-name: fallback-name-skill
-description: Test skill
----
-Test content
-"#;
-        fs::write(source_dir.join("SKILL.md"), skill_md).unwrap();
-
-        let manifest_content = r#"[tool.fastskill]
-skills_directory = ".claude/skills"
-
-[dependencies]
-"#;
-        fs::write(temp_dir.path().join("skill-project.toml"), manifest_content).unwrap();
-
-        let config = ServiceConfig {
-            skill_storage_path: skills_dir,
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-
-        let args = AddArgs {
-            source: source_dir.display().to_string(),
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: false,
-            group: None,
-            recursive: false,
-        };
-
-        let result = execute_add(&service, args, false).await;
-
-        assert!(result.is_ok(), "Add should succeed: {:?}", result);
-
-        let skill_id = fastskill_core::SkillId::new("fallback-name-skill".to_string()).unwrap();
-        let skill = service.skill_manager().get_skill(&skill_id).await.unwrap();
-        assert!(skill.is_some(), "Should use name from SKILL.md frontmatter");
-    }
-
-    #[tokio::test]
-    async fn test_add_skill_invalid_name_fails() {
-        let _lock = fastskill_core::test_utils::DIR_MUTEX
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let temp_dir = TempDir::new().unwrap();
-        let original_dir = std::env::current_dir().ok();
-
-        struct DirGuard(Option<std::path::PathBuf>);
-        impl Drop for DirGuard {
-            fn drop(&mut self) {
-                if let Some(dir) = &self.0 {
-                    let _ = std::env::set_current_dir(dir);
-                }
-            }
-        }
-        let _guard = DirGuard(original_dir);
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-
-        let skills_dir = temp_dir.path().join(".claude/skills");
-        fs::create_dir_all(&skills_dir).unwrap();
-
-        let source_dir = temp_dir.path().join("valid-directory-name");
-        fs::create_dir_all(&source_dir).unwrap();
-
-        let skill_md = r#"---
-name: INVALID-NAME
-description: Test skill
----
-Test content
-"#;
-        fs::write(source_dir.join("SKILL.md"), skill_md).unwrap();
-
-        let manifest_content = r#"[tool.fastskill]
-skills_directory = ".claude/skills"
-
-[dependencies]
-"#;
-        fs::write(temp_dir.path().join("skill-project.toml"), manifest_content).unwrap();
-
-        let config = ServiceConfig {
-            skill_storage_path: skills_dir,
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-
-        let args = AddArgs {
-            source: source_dir.display().to_string(),
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: false,
-            group: None,
-            recursive: false,
-        };
-
-        let result = execute_add(&service, args, false).await;
-
-        assert!(result.is_err(), "Should fail due to invalid name");
-    }
-
-    fn make_test_skill(dir: &std::path::Path) {
-        fs::create_dir_all(dir).unwrap();
-        let skill_content = r#"---
-name: editable-test-skill
-version: 1.0.0
-description: A test skill for editable mode testing
----
-
-# Editable Test Skill
-"#;
-        fs::write(dir.join("SKILL.md"), skill_content).unwrap();
-        let skill_project_content = r#"[metadata]
-id = "editable-test-skill"
-version = "1.0.0"
-"#;
-        fs::write(dir.join("skill-project.toml"), skill_project_content).unwrap();
-    }
-
-    #[tokio::test]
-    async fn test_add_editable_local_rejects_git_url() {
-        let temp_dir = TempDir::new().unwrap();
-        let config = ServiceConfig {
-            skill_storage_path: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-        let args = AddArgs {
-            source: "https://github.com/org/repo".to_string(),
-            source_type: Some("git".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: true,
-            group: None,
-            recursive: false,
-        };
-        let result = execute_add(&service, args, false).await;
-        assert!(
-            matches!(result, Err(CliError::Config(_))),
-            "Expected Config error, got: {:?}",
-            result
-        );
-        if let Err(CliError::Config(msg)) = result {
-            assert!(
-                msg.contains("only supported for local folder sources"),
-                "Error should mention local folder sources: {}",
-                msg
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_add_editable_local_rejects_zip() {
-        let temp_dir = TempDir::new().unwrap();
-        let config = ServiceConfig {
-            skill_storage_path: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-        let args = AddArgs {
-            source: "/some/path/skill.zip".to_string(),
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: true,
-            group: None,
-            recursive: false,
-        };
-        let result = execute_add(&service, args, false).await;
-        assert!(
-            matches!(result, Err(CliError::Config(_))),
-            "Expected Config error for zip with editable, got: {:?}",
-            result
-        );
-        if let Err(CliError::Config(msg)) = result {
-            assert!(
-                msg.contains("only supported for local folder sources"),
-                "Error should mention local folder sources: {}",
-                msg
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn test_add_editable_local_rejects_registry_id() {
-        let temp_dir = TempDir::new().unwrap();
-        let config = ServiceConfig {
-            skill_storage_path: temp_dir.path().to_path_buf(),
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-        let args = AddArgs {
-            source: "scope/some-skill".to_string(),
-            source_type: Some("registry".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: true,
-            group: None,
-            recursive: false,
-        };
-        let result = execute_add(&service, args, false).await;
-        assert!(
-            matches!(result, Err(CliError::Config(_))),
-            "Expected Config error for registry id with editable, got: {:?}",
-            result
-        );
-        if let Err(CliError::Config(msg)) = result {
-            assert!(
-                msg.contains("only supported for local folder sources"),
-                "Error should mention local folder sources: {}",
-                msg
-            );
-        }
-    }
-
-    #[cfg(unix)]
-    #[tokio::test]
-    async fn test_add_editable_local_creates_symlink() {
-        let _lock = fastskill_core::test_utils::DIR_MUTEX
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let temp_dir = TempDir::new().unwrap();
-        let original_dir = std::env::current_dir().ok();
-
-        struct DirGuard(Option<std::path::PathBuf>);
-        impl Drop for DirGuard {
-            fn drop(&mut self) {
-                if let Some(dir) = &self.0 {
-                    let _ = std::env::set_current_dir(dir);
-                }
-            }
-        }
-        let _guard = DirGuard(original_dir);
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-
-        let skills_dir = temp_dir.path().join(".claude/skills");
-        fs::create_dir_all(&skills_dir).unwrap();
-
-        let source_dir = temp_dir.path().join("editable-skill-source");
-        make_test_skill(&source_dir);
-
-        let manifest_content = r#"[tool.fastskill]
-skills_directory = ".claude/skills"
-
-[dependencies]
-"#;
-        fs::write(temp_dir.path().join("skill-project.toml"), manifest_content).unwrap();
-
-        let config = ServiceConfig {
-            skill_storage_path: skills_dir.clone(),
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-
-        let args = AddArgs {
-            source: source_dir.display().to_string(),
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: true,
-            group: None,
-            recursive: false,
-        };
-
-        let result = execute_add(&service, args, false).await;
-        assert!(result.is_ok(), "Editable add should succeed: {:?}", result);
-
-        let storage_path = skills_dir.join("editable-test-skill");
-        let metadata = fs::symlink_metadata(&storage_path).unwrap();
-        assert!(
-            metadata.file_type().is_symlink(),
-            "Storage path should be a symlink for editable install, but it is not"
-        );
-    }
-
-    #[tokio::test]
-    async fn test_add_non_editable_local_creates_directory() {
-        let _lock = fastskill_core::test_utils::DIR_MUTEX
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-
-        let temp_dir = TempDir::new().unwrap();
-        let original_dir = std::env::current_dir().ok();
-
-        struct DirGuard(Option<std::path::PathBuf>);
-        impl Drop for DirGuard {
-            fn drop(&mut self) {
-                if let Some(dir) = &self.0 {
-                    let _ = std::env::set_current_dir(dir);
-                }
-            }
-        }
-        let _guard = DirGuard(original_dir);
-        std::env::set_current_dir(temp_dir.path()).unwrap();
-
-        let skills_dir = temp_dir.path().join(".claude/skills");
-        fs::create_dir_all(&skills_dir).unwrap();
-
-        let source_dir = temp_dir.path().join("non-editable-skill-source");
-        make_test_skill(&source_dir);
-
-        let manifest_content = r#"[tool.fastskill]
-skills_directory = ".claude/skills"
-
-[dependencies]
-"#;
-        fs::write(temp_dir.path().join("skill-project.toml"), manifest_content).unwrap();
-
-        let config = ServiceConfig {
-            skill_storage_path: skills_dir.clone(),
-            ..Default::default()
-        };
-        let mut service = FastSkillService::new(config).await.unwrap();
-        service.initialize().await.unwrap();
-
-        let args = AddArgs {
-            source: source_dir.display().to_string(),
-            source_type: Some("local".to_string()),
-            branch: None,
-            tag: None,
-            force: false,
-            editable: false,
-            group: None,
-            recursive: false,
-        };
-
-        let result = execute_add(&service, args, false).await;
-        assert!(
-            result.is_ok(),
-            "Non-editable add should succeed: {:?}",
-            result
-        );
-
-        let storage_path = skills_dir.join("editable-test-skill");
-        let metadata = fs::symlink_metadata(&storage_path).unwrap();
-        assert!(
-            !metadata.file_type().is_symlink(),
-            "Storage path should be a real directory for non-editable install, not a symlink"
-        );
-        assert!(
-            storage_path.is_dir(),
-            "Storage path should be a real directory"
-        );
     }
 }
