+++ crates/fastskill-cli/src/main.rs
@@ -21,10 +21,164 @@ mod error;
 pub mod runtime_selector;
 mod utils;
 
-use cli_framework::prelude::AppBuilder;
+use cli_framework::prelude::{AppBuilder, Command};
 use context::{FsCtx, FsState};
 use std::path::PathBuf;
 use std::sync::Arc;
+
+/// Register a command into the AppBuilder, rebinding `$builder` on success.
+///
+/// Calling conventions (last token(s) select how the execute fn is invoked):
+///   (none)       — no service, `execute_fn(args)`
+///   `global`     — no service, `execute_fn(args, state.global)`
+///   `svc`        — service (&), `execute_fn(&svc, args)`
+///   `svc_global` — service (&), `execute_fn(&svc, args, state.global)`
+///   `svc_val`    — service (value), `execute_fn(svc, args)`
+macro_rules! register_cmd {
+    // No service, args only
+    ($builder:ident, $id:literal, $summary:literal, $syntax:expr, $category:expr,
+     $expose_mcp:literal, $args_type:ty, $state:expr,
+     $module:ident, $fn:ident) => {
+        let $builder = $builder.register_command({
+            let state = Arc::clone(&$state);
+            Command {
+                id: $id,
+                summary: $summary,
+                syntax: $syntax,
+                category: $category,
+                spec: None,
+                validator: None,
+                expose_mcp: $expose_mcp,
+                execute: Arc::new(move |_ctx, _args| {
+                    let state = Arc::clone(&state);
+                    Box::pin(async move {
+                        let raw = state.raw_remaining_args.clone();
+                        let args = parse_from_args::<$args_type>(&raw[1..])?;
+                        $module::$fn(args).await?;
+                        Ok(())
+                    })
+                }),
+            }
+        })?;
+    };
+    // No service, args + state.global
+    ($builder:ident, $id:literal, $summary:literal, $syntax:expr, $category:expr,
+     $expose_mcp:literal, $args_type:ty, $state:expr,
+     $module:ident, $fn:ident, global) => {
+        let $builder = $builder.register_command({
+            let state = Arc::clone(&$state);
+            Command {
+                id: $id,
+                summary: $summary,
+                syntax: $syntax,
+                category: $category,
+                spec: None,
+                validator: None,
+                expose_mcp: $expose_mcp,
+                execute: Arc::new(move |_ctx, _args| {
+                    let state = Arc::clone(&state);
+                    Box::pin(async move {
+                        let raw = state.raw_remaining_args.clone();
+                        let args = parse_from_args::<$args_type>(&raw[1..])?;
+                        $module::$fn(args, state.global).await?;
+                        Ok(())
+                    })
+                }),
+            }
+        })?;
+    };
+    // Service (&), args only
+    ($builder:ident, $id:literal, $summary:literal, $syntax:expr, $category:expr,
+     $expose_mcp:literal, $args_type:ty, $state:expr,
+     $module:ident, $fn:ident, svc) => {
+        let $builder = $builder.register_command({
+            let state = Arc::clone(&$state);
+            Command {
+                id: $id,
+                summary: $summary,
+                syntax: $syntax,
+                category: $category,
+                spec: None,
+                validator: None,
+                expose_mcp: $expose_mcp,
+                execute: Arc::new(move |_ctx, _args| {
+                    let state = Arc::clone(&state);
+                    Box::pin(async move {
+                        let svc = state.service().await?;
+                        let raw = state.raw_remaining_args.clone();
+                        let args = parse_from_args::<$args_type>(&raw[1..])?;
+                        $module::$fn(&svc, args).await?;
+                        Ok(())
+                    })
+                }),
+            }
+        })?;
+    };
+    // Service (&), args + state.global
+    ($builder:ident, $id:literal, $summary:literal, $syntax:expr, $category:expr,
+     $expose_mcp:literal, $args_type:ty, $state:expr,
+     $module:ident, $fn:ident, svc_global) => {
+        let $builder = $builder.register_command({
+            let state = Arc::clone(&$state);
+            Command {
+                id: $id,
+                summary: $summary,
+                syntax: $syntax,
+                category: $category,
+                spec: None,
+                validator: None,
+                expose_mcp: $expose_mcp,
+                execute: Arc::new(move |_ctx, _args| {
+                    let state = Arc::clone(&state);
+                    Box::pin(async move {
+                        let svc = state.service().await?;
+                        let raw = state.raw_remaining_args.clone();
+                        let args = parse_from_args::<$args_type>(&raw[1..])?;
+                        $module::$fn(&svc, args, state.global).await?;
+                        Ok(())
+                    })
+                }),
+            }
+        })?;
+    };
+    // Service (value), args only
+    ($builder:ident, $id:literal, $summary:literal, $syntax:expr, $category:expr,
+     $expose_mcp:literal, $args_type:ty, $state:expr,
+     $module:ident, $fn:ident, svc_val) => {
+        let $builder = $builder.register_command({
+            let state = Arc::clone(&$state);
+            Command {
+                id: $id,
+                summary: $summary,
+                syntax: $syntax,
+                category: $category,
+                spec: None,
+                validator: None,
+                expose_mcp: $expose_mcp,
+                execute: Arc::new(move |_ctx, _args| {
+                    let state = Arc::clone(&state);
+                    Box::pin(async move {
+                        let svc = state.service().await?;
+                        let raw = state.raw_remaining_args.clone();
+                        let args = parse_from_args::<$args_type>(&raw[1..])?;
+                        $module::$fn(svc, args).await?;
+                        Ok(())
+                    })
+                }),
+            }
+        })?;
+    };
+}
+
+fn or_exit<T, E: std::fmt::Display>(result: Result<T, E>, msg: &str) -> T {
+    match result {
+        Ok(v) => v,
+        Err(e) => {
+            eprintln!("{}: {}", msg, e);
+            std::process::exit(1);
+        }
+    }
+}
 
 use commands::{
     add, analyze, auth, disable, eval, init, install, list, marketplace, package, publish, read,
@@ -67,12 +221,14 @@ pub fn strip_global_flags(args: Vec<String>) -> (Option<PathBuf>, bool, bool, Ve
             verbose = true;
             i += 1;
         } else if arg == "--repositories-path" {
+            // Accepted for backward compatibility with older callers; value is intentionally ignored.
             if i + 1 < args.len() {
                 i += 2;
             } else {
                 i += 1;
             }
         } else if arg.starts_with("--repositories-path=") {
+            // Accepted for backward compatibility; value is intentionally ignored.
             i += 1;
         } else {
             remaining.push(arg.clone());
@@ -106,25 +262,18 @@ async fn main() {
     ));
     let ctx = FsCtx;
 
-    let builder = match build_app(AppBuilder::new(), Arc::clone(&state)) {
-        Ok(b) => b,
-        Err(e) => {
-            eprintln!("Error building app: {}", e);
-            std::process::exit(1);
-        }
-    };
+    let builder = or_exit(
+        build_app(AppBuilder::new(), Arc::clone(&state)),
+        "Error building app",
+    );
 
-    let mut app = match builder
-        .with_version("fastskill", fastskill_core::VERSION)
-        .with_git_sha_short(None)
-        .build(ctx)
-    {
-        Ok(a) => a,
-        Err(e) => {
-            eprintln!("Error initialising app: {}", e);
-            std::process::exit(1);
-        }
-    };
+    let mut app = or_exit(
+        builder
+            .with_version("fastskill", fastskill_core::VERSION)
+            .with_git_sha_short(None)
+            .build(ctx),
+        "Error initialising app",
+    );
 
     match app.run_with_args(remaining_args).await {
         Ok(()) => std::process::exit(0),
@@ -136,464 +285,274 @@ async fn main() {
 }
 
 fn build_app(builder: AppBuilder, state: Arc<FsState>) -> anyhow::Result<AppBuilder> {
-    use cli_framework::prelude::Command;
-
-    let builder = builder
-        // ── No-service commands ──────────────────────────────────────────────
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "init",
-                summary: "Initialize skill-project.toml in current skill directory",
-                syntax: Some("init [OPTIONS]"),
-                category: Some("project"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<init::InitArgs>(&raw[1..])?;
-                        init::execute_init(args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "install",
-                summary: "Apply manifest: install skills from skill-project.toml [dependencies]",
-                syntax: Some("install [OPTIONS]"),
-                category: Some("packages"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<install::InstallArgs>(&raw[1..])?;
-                        install::execute_install(args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "update",
-                summary: "Update skills to latest versions",
-                syntax: Some("update [SKILL_ID] [OPTIONS]"),
-                category: Some("packages"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<update::UpdateArgs>(&raw[1..])?;
-                        update::execute_update(args, state.global).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "show",
-                summary: "Show metadata for one or all installed skills",
-                syntax: Some("show [SKILL_ID] [OPTIONS]"),
-                category: Some("discovery"),
-                spec: None,
-                validator: None,
-                expose_mcp: true,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<show::ShowArgs>(&raw[1..])?;
-                        show::execute_show(args, state.global).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "package",
-                summary: "Package skills into ZIP artifacts with versioning",
-                syntax: Some("package [OPTIONS]"),
-                category: Some("publishing"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<package::PackageArgs>(&raw[1..])?;
-                        package::execute_package(args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "publish",
-                summary: "Publish artifacts to remote API or local folder",
-                syntax: Some("publish [OPTIONS]"),
-                category: Some("publishing"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<publish::PublishArgs>(&raw[1..])?;
-                        publish::execute_publish(args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        // ── Subcommand-tree commands (no service) ───────────────────────────
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "repos",
-                summary: "Manage repository list and browse remote skill catalog",
-                syntax: Some("repos <SUBCOMMAND>"),
-                category: Some("repositories"),
-                spec: None,
-                validator: None,
-                expose_mcp: true,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<repos::ReposArgs>(&raw[1..])?;
-                        repos::execute_repos(args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "marketplace",
-                summary: "Create and manage skill marketplace artifacts",
-                syntax: Some("marketplace <SUBCOMMAND>"),
-                category: Some("publishing"),
-                spec: None,
-                validator: None,
-                expose_mcp: true,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<marketplace::MarketplaceArgs>(&raw[1..])?;
-                        marketplace::execute_marketplace(args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "auth",
-                summary: "Manage authentication for registries",
-                syntax: Some("auth <login|logout|whoami>"),
-                category: Some("auth"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<auth::AuthArgs>(&raw[1..])?;
-                        auth::execute_auth(args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "eval",
-                summary: "Evaluation commands for skill quality assurance",
-                syntax: Some("eval <run|validate>"),
-                category: Some("quality"),
-                spec: None,
-                validator: None,
-                expose_mcp: true,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<eval::EvalCommand>(&raw[1..])?;
-                        eval::execute_eval(args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        // ── Service commands ─────────────────────────────────────────────────
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "add",
-                summary: "Add a skill (from local path, zip, git URL, or registry ID)",
-                syntax: Some("add <SOURCE> [OPTIONS]"),
-                category: Some("packages"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<add::AddArgs>(&raw[1..])?;
-                        add::execute_add(&svc, args, state.global).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "analyze",
-                summary: "Diagnostic and analysis commands",
-                syntax: Some("analyze <matrix|duplicates|cluster>"),
-                category: Some("analysis"),
-                spec: None,
-                validator: None,
-                expose_mcp: true,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<analyze::AnalyzeCommand>(&raw[1..])?;
-                        analyze::execute_analyze(&svc, args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "disable",
-                summary: "Disable skills by ID",
-                syntax: Some("disable <SKILL_ID>..."),
-                category: Some("packages"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<disable::DisableArgs>(&raw[1..])?;
-                        disable::execute_disable(&svc, args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "list",
-                summary: "List installed skills and reconcile with skill-project.toml",
-                syntax: Some("list [OPTIONS]"),
-                category: Some("discovery"),
-                spec: None,
-                validator: None,
-                expose_mcp: true,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<list::ListArgs>(&raw[1..])?;
-                        list::execute_list(&svc, args, state.global).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "read",
-                summary: "Stream skill documentation (SKILL.md content) to stdout",
-                syntax: Some("read <SKILL_ID>"),
-                category: Some("discovery"),
-                spec: None,
-                validator: None,
-                expose_mcp: true,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<read::ReadArgs>(&raw[1..])?;
-                        read::execute_read(svc, args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "sync",
-                summary: "Sync installed skills to the agent's metadata file",
-                syntax: Some("sync [OPTIONS]"),
-                category: Some("packages"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<sync::SyncArgs>(&raw[1..])?;
-                        sync::execute_sync(&svc, args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "reindex",
-                summary: "Reindex the vector index for semantic search",
-                syntax: Some("reindex [OPTIONS]"),
-                category: Some("packages"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<reindex::ReindexArgs>(&raw[1..])?;
-                        reindex::execute_reindex(&svc, args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "remove",
-                summary: "Uninstall skills (removes from manifest and local installation)",
-                syntax: Some("remove <SKILL_ID>... [OPTIONS]"),
-                category: Some("packages"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<remove::RemoveArgs>(&raw[1..])?;
-                        remove::execute_remove(&svc, args, state.global).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "resolve",
-                summary: "Resolve skills as machine-readable JSON with canonical paths",
-                syntax: Some("resolve <QUERY> [OPTIONS]"),
-                category: Some("discovery"),
-                spec: None,
-                validator: None,
-                expose_mcp: true,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<resolve::ResolveArgs>(&raw[1..])?;
-                        resolve::execute_resolve(&svc, args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "search",
-                summary: "Search skills by query with explicit scope flags",
-                syntax: Some("search <QUERY> [--local|--remote] [OPTIONS]"),
-                category: Some("discovery"),
-                spec: None,
-                validator: None,
-                expose_mcp: true,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<search::SearchArgs>(&raw[1..])?;
-                        search::execute_search(&svc, args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?
-        .register_command({
-            let state = Arc::clone(&state);
-            Command {
-                id: "serve",
-                summary: "Start the FastSkill HTTP API server",
-                syntax: Some("serve [OPTIONS]"),
-                category: Some("server"),
-                spec: None,
-                validator: None,
-                expose_mcp: false,
-                execute: Arc::new(move |_ctx, _args| {
-                    let state = Arc::clone(&state);
-                    Box::pin(async move {
-                        let svc = state.service().await?;
-                        let raw = state.raw_remaining_args.clone();
-                        let args = parse_from_args::<serve::ServeArgs>(&raw[1..])?;
-                        serve::execute_serve(svc, args).await?;
-                        Ok(())
-                    })
-                }),
-            }
-        })?;
+    // ── No-service commands ──────────────────────────────────────────────────
+    register_cmd!(
+        builder,
+        "init",
+        "Initialize skill-project.toml in current skill directory",
+        Some("init [OPTIONS]"),
+        Some("project"),
+        false,
+        init::InitArgs,
+        state,
+        init,
+        execute_init
+    );
+    register_cmd!(
+        builder,
+        "install",
+        "Apply manifest: install skills from skill-project.toml [dependencies]",
+        Some("install [OPTIONS]"),
+        Some("packages"),
+        false,
+        install::InstallArgs,
+        state,
+        install,
+        execute_install
+    );
+    register_cmd!(
+        builder,
+        "update",
+        "Update skills to latest versions",
+        Some("update [SKILL_ID] [OPTIONS]"),
+        Some("packages"),
+        false,
+        update::UpdateArgs,
+        state,
+        update,
+        execute_update,
+        global
+    );
+    register_cmd!(
+        builder,
+        "show",
+        "Show metadata for one or all installed skills",
+        Some("show [SKILL_ID] [OPTIONS]"),
+        Some("discovery"),
+        true,
+        show::ShowArgs,
+        state,
+        show,
+        execute_show,
+        global
+    );
+    register_cmd!(
+        builder,
+        "package",
+        "Package skills into ZIP artifacts with versioning",
+        Some("package [OPTIONS]"),
+        Some("publishing"),
+        false,
+        package::PackageArgs,
+        state,
+        package,
+        execute_package
+    );
+    register_cmd!(
+        builder,
+        "publish",
+        "Publish artifacts to remote API or local folder",
+        Some("publish [OPTIONS]"),
+        Some("publishing"),
+        false,
+        publish::PublishArgs,
+        state,
+        publish,
+        execute_publish
+    );
+    // ── Subcommand-tree commands (no service) ────────────────────────────────
+    register_cmd!(
+        builder,
+        "repos",
+        "Manage repository list and browse remote skill catalog",
+        Some("repos <SUBCOMMAND>"),
+        Some("repositories"),
+        true,
+        repos::ReposArgs,
+        state,
+        repos,
+        execute_repos
+    );
+    register_cmd!(
+        builder,
+        "marketplace",
+        "Create and manage skill marketplace artifacts",
+        Some("marketplace <SUBCOMMAND>"),
+        Some("publishing"),
+        true,
+        marketplace::MarketplaceArgs,
+        state,
+        marketplace,
+        execute_marketplace
+    );
+    register_cmd!(
+        builder,
+        "auth",
+        "Manage authentication for registries",
+        Some("auth <login|logout|whoami>"),
+        Some("auth"),
+        false,
+        auth::AuthArgs,
+        state,
+        auth,
+        execute_auth
+    );
+    register_cmd!(
+        builder,
+        "eval",
+        "Evaluation commands for skill quality assurance",
+        Some("eval <run|validate>"),
+        Some("quality"),
+        true,
+        eval::EvalCommand,
+        state,
+        eval,
+        execute_eval
+    );
+    // ── Service commands ─────────────────────────────────────────────────────
+    register_cmd!(
+        builder,
+        "add",
+        "Add a skill (from local path, zip, git URL, or registry ID)",
+        Some("add <SOURCE> [OPTIONS]"),
+        Some("packages"),
+        false,
+        add::AddArgs,
+        state,
+        add,
+        execute_add,
+        svc_global
+    );
+    register_cmd!(
+        builder,
+        "analyze",
+        "Diagnostic and analysis commands",
+        Some("analyze <matrix|duplicates|cluster>"),
+        Some("analysis"),
+        true,
+        analyze::AnalyzeCommand,
+        state,
+        analyze,
+        execute_analyze,
+        svc
+    );
+    register_cmd!(
+        builder,
+        "disable",
+        "Disable skills by ID",
+        Some("disable <SKILL_ID>..."),
+        Some("packages"),
+        false,
+        disable::DisableArgs,
+        state,
+        disable,
+        execute_disable,
+        svc
+    );
+    register_cmd!(
+        builder,
+        "list",
+        "List installed skills and reconcile with skill-project.toml",
+        Some("list [OPTIONS]"),
+        Some("discovery"),
+        true,
+        list::ListArgs,
+        state,
+        list,
+        execute_list,
+        svc_global
+    );
+    register_cmd!(
+        builder,
+        "read",
+        "Stream skill documentation (SKILL.md content) to stdout",
+        Some("read <SKILL_ID>"),
+        Some("discovery"),
+        true,
+        read::ReadArgs,
+        state,
+        read,
+        execute_read,
+        svc_val
+    );
+    register_cmd!(
+        builder,
+        "sync",
+        "Sync installed skills to the agent's metadata file",
+        Some("sync [OPTIONS]"),
+        Some("packages"),
+        false,
+        sync::SyncArgs,
+        state,
+        sync,
+        execute_sync,
+        svc
+    );
+    register_cmd!(
+        builder,
+        "reindex",
+        "Reindex the vector index for semantic search",
+        Some("reindex [OPTIONS]"),
+        Some("packages"),
+        false,
+        reindex::ReindexArgs,
+        state,
+        reindex,
+        execute_reindex,
+        svc
+    );
+    register_cmd!(
+        builder,
+        "remove",
+        "Uninstall skills (removes from manifest and local installation)",
+        Some("remove <SKILL_ID>... [OPTIONS]"),
+        Some("packages"),
+        false,
+        remove::RemoveArgs,
+        state,
+        remove,
+        execute_remove,
+        svc_global
+    );
+    register_cmd!(
+        builder,
+        "resolve",
+        "Resolve skills as machine-readable JSON with canonical paths",
+        Some("resolve <QUERY> [OPTIONS]"),
+        Some("discovery"),
+        true,
+        resolve::ResolveArgs,
+        state,
+        resolve,
+        execute_resolve,
+        svc
+    );
+    register_cmd!(
+        builder,
+        "search",
+        "Search skills by query with explicit scope flags",
+        Some("search <QUERY> [--local|--remote] [OPTIONS]"),
+        Some("discovery"),
+        true,
+        search::SearchArgs,
+        state,
+        search,
+        execute_search,
+        svc
+    );
+    register_cmd!(
+        builder,
+        "serve",
+        "Start the FastSkill HTTP API server",
+        Some("serve [OPTIONS]"),
+        Some("server"),
+        false,
+        serve::ServeArgs,
+        state,
+        serve,
+        execute_serve,
+        svc_val
+    );
 
     Ok(builder)
 }
